Added boundaries for when we're dragging items.
Boundaries for items dragged. Next up working on the collision with the bottom UI elements to complete the level. Also found a bug where if the item is dragged into the bottom combo area, you can't drag it until something else is there. Will have to look into that before finalizing the dragging.