Drag and drop is almost fully completed. Snapping to the combo box works.
If you hold the item between 2 boxes the size gets messed up. Other than that, dragging and dropping is completed. Just need to make the item appear in the result box and ensure that no combos can be made until that box is empty. Then level progression will be implemented.