Halfway there on the scaling/transform fix.
Background items will always snap back to their original positions. May have to change this when we're looking for stuff hidden behind them (That or just have the item hidden pop out). Also scaling is coming along to be more dynamic based on the size of the object. Next up is making sure that the items can't be put in weird spots on the map (May just have everything snap back to place tbh)