Bug fixes for combo area
Fixed the bug where if you dragged an item between both combo boxes it would auto-combine into the comboItem. If we need 2 of the same item to combine into something however, this code will have to be refactored.

Also, if an item is in the result box when 2 items are trying to be combined (And it's a combo item) , the combo items now have initial positions in their scriptable object boxes that can be manipulated in the inspector.