Added main menu and pause menu
Main menu and stage selected added. We can discuss how we want it to look like in the final version, this is just a general idea for demo purposes.

Also, added the pause menu. I have an idea of what to do with this but we can also discuss if it would be applicable to casual audiences.